+++ src/components/QuestionList/index.jsx
@@ -1,5 +1,5 @@
 import QuestionAndAnswer from './QuestionAndAnswer';
-import styles from './QuestionListItem.module.css';
+import styles from './QuestionList.module.css';
 
 // # API 함수 props에서 제거
 // # subject는 새롭게 요청하기 보단 상위 컴포넌트에서 받아온다.
@@ -15,7 +15,7 @@ import styles from './QuestionListItem.module.css';
  * @param {function} props.onUpdate question 리스트 업데이트 핸들러
  * @returns {React.JSX} 질문 리스트 아이템 컴포넌트
  */
-function QuestionListItem({ isSubjectOwner, subject, questions, onUpdate }) {
+function QuestionList({ isSubjectOwner, subject, questions, onUpdate }) {
   // # subject를 직접 받아오니 id를 url param에서 얻을 필요가 없음
   // # 상태값과 이를 정의하기 위한 useEffet 제거
 
@@ -38,4 +38,4 @@ function QuestionListItem({ isSubjectOwner, subject, questions, onUpdate }) {
   );
 }
 
-export default QuestionListItem;
+export default QuestionList;
